Refactor some Commands
