+++ src/main/java/frc/robot/commands/auto/TurnToTarget.java
@@ -5,39 +5,45 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.auto;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class GrabPanel extends Command {
-  public GrabPanel() {
+/**
+ * A command to turn the robot towards the nearest vision target
+ */
+public class TurnToTarget extends Command {
+  /**
+   * A command to turn the robot towards the nearest vision target
+   */
+  public TurnToTarget() {
     // Use requires() here to declare subsystem dependencies
     // eg. requires(chassis);
-    requires(Robot.hatchClaw);
+    requires(Robot.drivetrain);
   }
-
+  
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
   }
-
+  
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
   }
-
+  
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
     return false;
   }
-
+  
   // Called once after isFinished returns true
   @Override
   protected void end() {
   }
-
+  
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
